requirements.txt
line-bot-sdk==3.7.0
flask==3.0.0
azure-ai-language-conversations